==> EVC_Nov23/03_DuplicatedIndividualsAcrossStudies.R <==
### in this script duplicated data is identified. As the same data can be present partially or entirely in multiple studies in MB, these need to be identified when downloading all data of one or multiple species. Given that often the individual_local_identifier is different for exactly the same data, we rely heavily on tag id, as this is more likely to not change. Here we find duplicated individuals/tags within/across studies:
# - check duplicated indv-tag-sps
# - check duplicated by tag-sps
# - check individuals with multiple tags simultaneously
# code checks if possible duplicated indiv overlap in tracking time
# OPTION: keep the one with longest duration or with most gps points ~L105

############# function ####################################
library('move2')
library('lubridate')
library("data.table")

## get a reference table with 1 line per individualID-tagID combination
# So individuals which have multiple tags/deployments will have one line per tagID, associated to the start and end of the tracking time for that specific tag.
# This is important for finding duplicated tagIDs within or across studies.

# path_to_indv_move2 <- "/home/ascharf/Documents/Projects/Drylands/EVC23/2.vultureIndv_mv2_clean_empty_duply//2034212088_G32753 - turquoise black 25 - 3633 - rehab.rds"

referenceTableStudies <-  function(path_to_indv_move2){ 
  print(path_to_indv_move2) # this makes it easy to find which one gave an error
  indiv_mv2 <- readRDS(path_to_indv_move2)
  indiv_mv2_td <- mt_track_data(indiv_mv2)
  fileNameL <- strsplit(path_to_indv_move2,"/")
  fileName <- fileNameL[[1]][length(fileNameL[[1]])]
  
  # this is to account for when individual_local_identifier is missing
  if(all(is.na(indiv_mv2_td$individual_local_identifier))==T
     & all(is.na(indiv_mv2_td$tag_local_identifier))==F){ 
    indiv_mv2 <- mutate_track_data(indiv_mv2, individual_local_identifier = indiv_mv2_td$tag_local_identifier)
  }else if(all(is.na(indiv_mv2_td$individual_local_identifier))==T
           & all(is.na(indiv_mv2_td$tag_local_identifier))==T){warning("This study has neither individual nor tag identifier, study excluded from this step.")}
  # this is to account for when tag_local_identifier is missing
  if(all(is.null(indiv_mv2_td$tag_local_identifier))==T){ 
    indiv_mv2 <- mutate_track_data(indiv_mv2, tag_local_identifier=indiv_mv2_td$tag_id)
  }
  
  # this is to account for individuals with multiple tags (i.e deployments)
  tagList <- split(indiv_mv2, mt_track_id(indiv_mv2))
  
  indTable <- as.data.frame(rbindlist(lapply(tagList, function(tag){
    # tag <- tagList[[1]]
    tag_td <- mt_track_data(tag)
    rdf <- data.frame(
      fileName = fileName,
      MBid = tag_td$study_id,
      individual_local_identifier = as.character(tag_td$individual_local_identifier),
      tag_local_identifier = as.character(tag_td$tag_local_identifier),
      # species = as.character(tag_td$taxon_canonical_name),
      species = if(!is.null(tag_td$taxon_canonical_name)){tag_td$taxon_canonical_name}else{tag_td$taxon_ids},
      animal_life_stage = if(is.null(tag_td$animal_life_stage)){NA}else if(is.na(unique(tag_td$animal_life_stage))){NA}else{unique(as.character(tag_td$animal_life_stage))}, ## included not for filtering here, but including it as needed afterwards
      tracking_duration_days = round(as.numeric(difftime(range(mt_time(tag))[2],range(mt_time(tag))[1],units="days")),2),
      tracking_start_date = range(mt_time(tag))[1],  # Important to keep the full timestamp, as in the same day a tag can be removed from one individual and put on another one.
      tracking_end_date = range(mt_time(tag))[2],
      GPSpts_total = nrow(tag),
      median_timelag_mins = median(mt_time_lags(tag),na.rm=T),
      min_timelag_mins = min(mt_time_lags(tag),na.rm=T)
    )
    units(rdf$median_timelag_mins) <- units::make_units(minute)
    units(rdf$min_timelag_mins) <- units::make_units(minute)
    units(rdf$tracking_duration_days) <- units::make_units(day)
    rdf$median_timelag_mins <- round(rdf$median_timelag_mins)
    rdf$min_timelag_mins <- round(rdf$min_timelag_mins)
    return(rdf)
  })))
  return(indTable)
}

########### end function #################################

#____________________________________
## Individuals reference table: ####
# Create a reference table, one entry per individual: #####

genPath <- "/home/ascharf/Documents/Projects/Drylands/EVC23/"
pthClean <- "2.vultureIndv_mv2_clean_empty_duply/"
dataPath <- paste0(genPath,pthClean)

fls <- list.files(dataPath, pattern="rds", full.names = T)

start_time<- Sys.time()
referenceTableStudies_ALL <- as.data.frame(rbindlist(lapply(fls, referenceTableStudies)))
end_time <- Sys.time() 
end_time-start_time # ~20mins

saveRDS(referenceTableStudies_ALL, file=paste0(genPath,"/referenceTableStudies_ALL_original.rds")) ## just to making sure to have a copy that is untouched, as after this it will be modified and overwritten....  

# length(unique(referenceTableStudies_ALL$MBid))==length(fls)
# ids <- sapply(strsplit(fls,"_"), "[[", 2)
# fls[!ids %in% unique(referenceTableStudies_ALL$MBid)]

#_______________________________________________________________
## Find duplicated individuals/tags within/across studies: ####
# this script might have to be adjusted depending on the particular issues that come up when all studies are gathered

library(DescTools) # for function %overlaps%
library(dplyr)
library(bit64)
referenceTableStudies_ALL <-  readRDS(paste0(genPath,"/referenceTableStudies_ALL_original.rds"))
referenceTableStudies_ALL$MBid <- as.character(referenceTableStudies_ALL$MBid) ## class integer64 sometimes gets in loops, and converts to something else (very long number)

## ensure that max tracking end date corresponds to download date.
max(referenceTableStudies_ALL$tracking_end_date) 

## choose which duplicate to keep, with longest duration ("duration") or with largest number of gps points ("locations")
keepCriterion <- "duration" # "duration", "locations"

## adding some columns for managing the table
referenceTableStudies_ALL$rowID <- paste0("rID_",1:nrow(referenceTableStudies_ALL))
referenceTableStudies_ALL$excluded <- "no"
referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps <- NA
referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps2 <- NA


## adding columns by which to filter
referenceTableStudies_ALL$Ind_Tag_Sps <- paste0(referenceTableStudies_ALL$individual_local_identifier,"_",referenceTableStudies_ALL$tag_local_identifier,"_",referenceTableStudies_ALL$species)
referenceTableStudies_ALL$Tag_Sps <- paste0(referenceTableStudies_ALL$tag_local_identifier,"_",referenceTableStudies_ALL$species)
referenceTableStudies_ALL$Stu_Ind_Sps <- paste0(referenceTableStudies_ALL$MBid,"_",referenceTableStudies_ALL$individual_local_identifier,"_",referenceTableStudies_ALL$species)

#### FOR THE STORKS THE RESULTS WERE BETTER WHEN ONLY USING THE TAG_SPS OVERLAP CRITERION AS NAMES CHANGE A LOT, ~L122-191 were not run ####

# ## check duplicated indv-tag-sps ----
# # get table containing all rows of duplication
# dupliTab_ITS <- referenceTableStudies_ALL[referenceTableStudies_ALL$Ind_Tag_Sps %in% referenceTableStudies_ALL$Ind_Tag_Sps[duplicated(referenceTableStudies_ALL$Ind_Tag_Sps)],]
# 
# print(paste0(nrow(dupliTab_ITS), " rows duplicated by indiv, tag and sps"))
# if(nrow(dupliTab_ITS)>1){
#   # splitting to select which of the duplicateds to keep
#   dupliTab_ITS_l <- split(dupliTab_ITS, dupliTab_ITS$Ind_Tag_Sps)
#   # toExclude_l <- lapply(dupliTab_ITS_l, function(x){
#   for(i in 1:length(dupliTab_ITS_l)){
#     x <- dupliTab_ITS_l[[i]]
#     # make matrix to find all paired combinations
#     combiMatrix <- expand.grid(x$rowID,x$rowID,stringsAsFactors=F)
#     colnames(combiMatrix) <- c("id1", "id2")
#     combiMatrix <- combiMatrix[combiMatrix$id1!=combiMatrix$id2,] # removing comparisons to them selves (eg. ID2, ID2)
#     combiMatrix <- combiMatrix[duplicated(t(apply(combiMatrix, 1, sort))),] # removing same comparison, but different order (eg. ID3-ID4 and ID4-ID3) as %overlaps% does not care about order
#     combiMatrix$overlap <- FALSE
#     # checking if overlap is true for all pairs
#     for(i in 1:(nrow(combiMatrix))){
#       if(c(x$tracking_start_date[x$rowID==combiMatrix$id1[i]], x$tracking_end_date[x$rowID==combiMatrix$id1[i]]) %overlaps% c(x$tracking_start_date[x$rowID==combiMatrix$id2[i]], x$tracking_end_date[x$rowID==combiMatrix$id2[i]]) == TRUE){
#         combiMatrix[i,"overlap"] <- TRUE
#       }
#     }
#     # not overlapping duplicated tag_sps in time: probably same(or different) study deployed tag on different individual
#     overlapping <- combiMatrix[combiMatrix$overlap==TRUE,]
#     # accounting for table having multiple rows, e.g. 2 individuals got tagged with same tag, both are duplicated
#     if(nrow(overlapping) > 0){
#       ## keeping the one with the most gps pts
#       if(keepCriterion=="locations"){
#         print("keeping duplicate with largest number of locations")
#         exclude_l <- lapply(1:(nrow(overlapping)),function(i){
#           rowToExclude <- which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "GPSpts_total"])  # if same number of gps pts, 1st one gets declared as min
#           x$rowID[x$rowID %in% overlapping[i,c("id1","id2")]][rowToExclude] # this works also when overlapping has multiple rows
#         })
#       }
#       ## keeping the one with the longest duration
#       if(keepCriterion=="duration"){
#         print("keeping duplicate with longest duration")
#         exclude_l <- lapply(1:(nrow(overlapping)),function(i){
#           # if both tags have same tracking duration we take the one with more GPS locations, otherwise the one with longer duration
#           if(x[x$rowID %in% overlapping[i,"id1"], "tracking_duration_days"] == x[x$rowID %in% overlapping[i,"id2"], "tracking_duration_days"]){
#             x$rowID[which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "GPSpts_total"])]
#           }else{x$rowID[which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "tracking_duration_days"])] # if same number of gps pts, 1st one gets declared as min
#           }
#         })
#       }
#       ## PROBLEM, getting all duplicates together, sometimes its 2 and 2...
#       exclude <- unique(unlist(exclude_l))
#       # return(exclude)
#       referenceTableStudies_ALL$excluded[referenceTableStudies_ALL$rowID %in% exclude] <- "yes_duplicated_Ind_Tag_Sps"
#       ## indicating for the group of duplicated, which of them is kept, to afterwards better evaluate which are true duplicates
#       all <- unique(c(overlapping$id1,overlapping$id2))
#       keep <- all[!all%in%exclude]
#       if(length(keep)==1){indKeep <- paste0(x$MBid[x$rowID==keep],"_", x$Ind_Tag_Sps[x$rowID==keep])
#       print(indKeep)
#       referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps2[referenceTableStudies_ALL$rowID %in% all] <- indKeep
#       }else{
#         for(j in 1:length(keep)){
#           sel <- overlapping%>%filter_all(any_vars(. %in% c(keep[j])))
#           indKeep <- paste0(x$MBid[x$rowID==keep[j]],"_", x$Ind_Tag_Sps[x$rowID==keep[j]])
#           print(indKeep)
#           referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps2[referenceTableStudies_ALL$rowID %in% c(sel[,1],sel[,2])] <- indKeep
#         }
#       }
#     }
#   }
# }
# 

## check duplicated by tag-sps ----
# get table containing all rows of duplication
dupliTab_TS <- referenceTableStudies_ALL[referenceTableStudies_ALL$excluded=="no",] ## removing those that have been already identified by duplicated Ind_Tag_Sps
dupliTab_TS <- dupliTab_TS[dupliTab_TS$Tag_Sps %in% dupliTab_TS$Tag_Sps[duplicated(dupliTab_TS$Tag_Sps)],] 

print(paste0(nrow(dupliTab_TS), " rows duplicated by tag and sps"))
if(nrow(dupliTab_TS)>1){
  # splitting to select which of the duplicateds to keep
  dupliTab_TS_l <- split(dupliTab_TS, dupliTab_TS$Tag_Sps)
  # toExclude_l <- lapply(dupliTab_TS_l, function(x){
  for(i in 1:length(dupliTab_TS_l)){
    print(i)
    x <- dupliTab_TS_l[[i]]
    
    # make matrix to find all paired combinations
    combiMatrix <- expand.grid(x$rowID,x$rowID,stringsAsFactors=F)
    colnames(combiMatrix) <- c("id1", "id2")
    combiMatrix <- combiMatrix[combiMatrix$id1!=combiMatrix$id2,] # removing comparisons to them selves (eg. ID2, ID2)
    combiMatrix <- combiMatrix[duplicated(t(apply(combiMatrix, 1, sort))),] # removing same comparison, but different order (eg. ID3-ID4 and ID4-ID3) as %overlaps% does not care about order
    combiMatrix$overlap <- FALSE
    # checking if overlap is true for all pairs
    for(i in 1:(nrow(combiMatrix))){
      if(c(x$tracking_start_date[x$rowID==combiMatrix$id1[i]], x$tracking_end_date[x$rowID==combiMatrix$id1[i]]) %overlaps% c(x$tracking_start_date[x$rowID==combiMatrix$id2[i]], x$tracking_end_date[x$rowID==combiMatrix$id2[i]]) == TRUE){
        combiMatrix[i,"overlap"] <- TRUE
      }
    }
    # not overlapping duplicated tag_sps in time: probably same(or different) study deployed tag on different individual
    overlapping <- combiMatrix[combiMatrix$overlap==TRUE,] 
    # accounting for table having multiple rows, e.g. 2 individuals got tagged with same tag, both are duplicated
    if(nrow(overlapping) > 0){
      ## keeping the one with the most gps pts
      if(keepCriterion=="locations"){
        print("keeping duplicate with largest number of locations")
        exclude_l <- lapply(1:(nrow(overlapping)),function(i){
          rowToExclude <- which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "GPSpts_total"])  # if same number of gps pts, 1st one gets declared as min
          x$rowID[x$rowID %in% overlapping[i,c("id1","id2")]][rowToExclude] # this works also when overlapping has multiple rows
        })
      }
      ## keeping the one with the longest duration
      if(keepCriterion=="duration"){
        print("keeping duplicate with longest duration")
        exclude_l <- lapply(1:(nrow(overlapping)),function(i){
          # if both tags have same tracking duration we take the one with more GPS locations, otherwise the one with longer duration
          if(x[x$rowID %in% overlapping[i,"id1"], "tracking_duration_days"] == x[x$rowID %in% overlapping[i,"id2"], "tracking_duration_days"]){
            x$rowID[which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "GPSpts_total"])]
          }else{x$rowID[which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "tracking_duration_days"])] # if same number of gps pts, 1st one gets declared as min
          }
        })
      }
      ## PROBLEM, getting all duplicates together, sometimes its 2 and 2...
      exclude <- unique(unlist(exclude_l))
      # return(exclude)
      referenceTableStudies_ALL$excluded[referenceTableStudies_ALL$rowID %in% exclude] <- "yes_duplicated_Tag_Sps"
      ## indicating for the group of duplicated, which of them is kept, to afterwards better evaluate which are true duplicates
      all <- unique(c(overlapping$id1,overlapping$id2))
      keep <- all[!all%in%exclude]
      if(length(keep)==1){indKeep <- paste0(x$MBid[x$rowID==keep],"_", x$Ind_Tag_Sps[x$rowID==keep])
      print(indKeep)
      referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps2[referenceTableStudies_ALL$rowID %in% all] <- indKeep
      }else{
        for(j in 1:length(keep)){
          sel <- overlapping%>%filter_all(any_vars(. %in% c(keep[j])))
          indKeep <- paste0(x$MBid[x$rowID==keep[j]],"_", x$Ind_Tag_Sps[x$rowID==keep[j]])
          print(indKeep)
          referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps2[referenceTableStudies_ALL$rowID %in% c(sel[,1],sel[,2])] <- indKeep
        }                  
      }
    }
  }
}

## checking if a selected indiv got overwritten by another one
hck <- referenceTableStudies_ALL[, c("kept_MB_Ind_Tag_Sps","kept_MB_Ind_Tag_Sps2")]
# hck <- hck[rowSums(is.na(hck)) != ncol(hck), ]
hck <- hck[complete.cases(hck),]
identical(hck$kept_MB_Ind_Tag_Sps,hck$kept_MB_Ind_Tag_Sps2)

referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps[is.na(referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps)] <- referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps2[is.na(referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps)]
referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps2 <- NULL

saveRDS(referenceTableStudies_ALL, file=paste0(genPath,"/referenceTableStudies_ALL_excludedColumn.rds"))


## check individuals with multiple tags simultaneously ----
# get table containing all rows of duplication
dupliTab_I <- referenceTableStudies_ALL[referenceTableStudies_ALL$excluded=="no",] ## removing those that have been already identified by duplicated Tag_Sps
dupliTab_I <- dupliTab_I[dupliTab_I$Stu_Ind_Sps %in% dupliTab_I$Stu_Ind_Sps[duplicated(dupliTab_I$Stu_Ind_Sps)],] 

print(paste0(nrow(dupliTab_I), " rows duplicated by individual"))
if(nrow(dupliTab_I)>1){
  # splitting to select which of the duplicateds to keep
  dupliTab_I_l <- split(dupliTab_I, dupliTab_I$Stu_Ind_Sps)
  toExclude_l <- lapply(dupliTab_I_l, function(x){
    # make matrix to find all paired combinations
    combiMatrix <- expand.grid(x$rowID,x$rowID,stringsAsFactors=F)
    colnames(combiMatrix) <- c("id1", "id2")
    combiMatrix <- combiMatrix[combiMatrix$id1!=combiMatrix$id2,] # removing comparisons to them selves (eg. ID2, ID2)
    combiMatrix <- combiMatrix[duplicated(t(apply(combiMatrix, 1, sort))),] # removing same comparison, but different order (eg. ID3-ID4 and ID4-ID3) as %overlaps% does not care about order
    combiMatrix$overlap <- FALSE
    # checking if overlap is true for all pairs
    for(i in 1:(nrow(combiMatrix))){
      if(c(x$tracking_start_date[x$rowID==combiMatrix$id1[i]], x$tracking_end_date[x$rowID==combiMatrix$id1[i]]) %overlaps% c(x$tracking_start_date[x$rowID==combiMatrix$id2[i]], x$tracking_end_date[x$rowID==combiMatrix$id2[i]]) == TRUE){
        combiMatrix[i,"overlap"] <- TRUE
      }
    }
    overlapping <- combiMatrix[combiMatrix$overlap==TRUE,] 
    if(nrow(overlapping) > 0){
      ## keeping the one with the most gps pts
      if(keepCriterion=="locations"){
        print("keeping duplicate with largest number of locations")
        exclude_l <- lapply(1:(nrow(overlapping)),function(i){
          rowToExclude <- which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "GPSpts_total"])  # if same number of gps pts, 1st one gets declared as min
          x$rowID[x$rowID %in% overlapping[i,c("id1","id2")]][rowToExclude] # this works also when overlapping has multiple rows
        })
      }
      ## keeping the one with the longest duration
      if(keepCriterion=="duration"){
        print("keeping duplicate with longest duration")
        exclude_l <- lapply(1:(nrow(overlapping)),function(i){
          # if both tags have same tracking duration we take the one with more GPS locations, otherwise the one with longer duration
          if(x[x$rowID %in% overlapping[i,"id1"], "tracking_duration_days"] == x[x$rowID %in% overlapping[i,"id2"], "tracking_duration_days"]){
            x$rowID[which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "GPSpts_total"])]
          }else{x$rowID[which.min(x[x$rowID %in% overlapping[i,c("id1","id2")], "tracking_duration_days"])] # if same number of gps pts, 1st one gets declared as min
          }
        })
      }
      exclude <- unlist(exclude_l)
      return(exclude)
    }
  })
  toExclude <- unlist(toExclude_l)
  referenceTableStudies_ALL$excluded[referenceTableStudies_ALL$rowID %in% toExclude] <- "yes_duplicated_Study-Individual-Species"
}

saveRDS(referenceTableStudies_ALL, file=paste0(genPath,"/referenceTableStudies_ALL_excludedColumn.rds"))

table(referenceTableStudies_ALL$excluded)

### ensuring all duplicates have been actually removed, and the correct one has been selected
## sometimes not all duplicates get identified, ie not all get excluded but one, probably given that there are 3 separate searches, not all pairs get compaired to all pairs
referenceTableStudies_ALL <- readRDS(file=paste0(genPath,"/referenceTableStudies_ALL_excludedColumn.rds"))

checkDF <- referenceTableStudies_ALL[complete.cases(referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps),]

checkL <- split(checkDF,checkDF$kept_MB_Ind_Tag_Sps)
table(unlist(lapply(checkL,nrow)))
checkL[[25]]

## sometimes 2 or more indiv are kept 
table(unlist(lapply(checkL,function(x){
  nrow(x[x$excluded=="no",])
})))

## retaining duplicate with max "tracking_duration_days" or "GPSpts_total" depending on choice of "keepCriterion" at the top of the script
for(i in 1:length(checkL)){
  grp <- checkL[[i]]
  # if(nrow(grp[grp$excluded=="no",])>1){ # it seems sometimes the wrong one got selected (i.e. shortest), not sure why.....
  
  if(keepCriterion=="locations"){
    kp <- grp$rowID[grp$GPSpts_total==max(grp$GPSpts_total)]
    if(length(kp)>1){kp <- kp[1]}
    rids <- grp$rowID
    rids <- rids[!rids==kp]
    referenceTableStudies_ALL$excluded[referenceTableStudies_ALL$rowID%in%kp] <- "no"
    referenceTableStudies_ALL$excluded[referenceTableStudies_ALL$rowID%in%rids] <- "yes_duplicated"
  }  
  
  if(keepCriterion=="duration"){
    kp <- grp$rowID[grp$tracking_duration_days==max(grp$tracking_duration_days)]
    if(length(kp)>1){kp <- kp[1]}
    rids <- grp$rowID
    rids <- rids[!rids==kp]
    referenceTableStudies_ALL$excluded[referenceTableStudies_ALL$rowID%in%kp] <- "no"
    referenceTableStudies_ALL$excluded[referenceTableStudies_ALL$rowID%in%rids] <- "yes_duplicated"
  }
  # }
}
# sanity check
checkDF <- referenceTableStudies_ALL[complete.cases(referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps),]
checkL <- split(checkDF,checkDF$kept_MB_Ind_Tag_Sps)
table(unlist(lapply(checkL,function(x){
  nrow(x[x$excluded=="no",])
})))

saveRDS(referenceTableStudies_ALL, file=paste0(genPath,"/referenceTableStudies_ALL_excludedColumn.rds"))

#_____________________
## Sanity checks 1 ####
# Plot the duplicated tags to make sure it's the same individual
library(move2)
library(ggplot2)

genPath <- "/home/ascharf/Documents/Projects/Drylands/EVC23/"
pthClean <- "2.vultureIndv_mv2_clean_empty_duply/"
dataPath <- paste0(genPath,pthClean)

referenceTableStudies_ALL <- readRDS(file=paste0(genPath,"/referenceTableStudies_ALL_excludedColumn.rds"))

checkDF <- referenceTableStudies_ALL[complete.cases(referenceTableStudies_ALL$kept_MB_Ind_Tag_Sps),]

checkL <- split(checkDF,checkDF$kept_MB_Ind_Tag_Sps)

pdf(file=paste0(genPath,"duplicatePlots.pdf"), width=20, height=12)
lapply(checkL, function(y){
  df <- data.frame(fileName=y$fileName, keep=y$excluded)
  df$keep[!df$keep=="no"] <- "duplicate"
  df$keep[df$keep=="no"] <- "keep"
  
  dupL <- lapply(y$fileName, function(x){
    m <- readRDS(paste0(dataPath,x))
    m$fileName <- x
    m$fileName <- as.factor(m$fileName)
    m$keep <- df$keep[df$fileName==x]
    mt_track_id(m) <- "fileName"
    return(m)
  })
  mv2 <- mt_stack(dupL)
  mv2 <- mt_as_track_attribute(mv2,"keep",.keep=T)
  mv2 <- mt_filter_per_interval(mv2,criterion="first",unit="day") # making plots lighter
  ggplot()+geom_sf(data=mt_track_lines(mv2), aes(color=keep))+facet_wrap(~fileName,nrow=1)
  # print(pl)
})
dev.off()

## this plot seemed empty, but there are only so few point that they cannot be seen
ind <- "150450794_30470469.rds"
m <- readRDS(paste0(dataPath,ind))
m <- mt_filter_per_interval(m,criterion="first",unit="day")
ggplot()+geom_sf(data=mt_track_lines(m))


ind <- "332044860_Mirabell - DER AN910 (eobs 3907).rds"
list.files(dataPath, pattern="332044860_Mirabell")
m <- readRDS(paste0(dataPath,ind)) 
# study,ind,tag,sps
kp <- paste0(mt_track_data(m)$study_id,"_",mt_track_data(m)$individual_local_identifier,"_",mt_track_data(m)$tag_local_identifier,"_",mt_track_data(m)$taxon_canonical_name)
checkL[[kp]]

#_____________________
## Sanity checks 2 ####
# Plot different tags on the same individual to make sure they overlap and are actual duplicates at least in part

# # fls <- list.files("MovementData/RawData", full.names = T)
# 
# table(referenceTableStudies_ALL$excluded)
# (dups <- as.character(referenceTableStudies_ALL$Stu_Ind_Sps[referenceTableStudies_ALL$excluded == "yes_duplicated_Study-Individual-Species"]))
# check <- referenceTableStudies_ALL[referenceTableStudies_ALL$Stu_Ind_Sps %in% dups,]
# check[,c("individual_local_identifier","tracking_start_date","tracking_end_date")]
# 
# studyIDs <- sapply(strsplit(dups, "_"), "[",1)
# 
# #i=136,137 "7023813_2049_Columba livia","7023813_2081_Columba livia" (tracks overlap but look different)
# dups
# i=152
# check[check$Stu_Ind_Sps==dups[i],c("individual_local_identifier","tracking_start_date","tracking_end_date","GPSpts_total","excluded")]
# 
# gps <- readRDS(grep(studyIDs[i], fls, value=T))
# 
# gps$Stu_Ind_Sps <- paste0(gps$study.id,"_",gps$individual_local_identifier,"_",gps$individual.taxon.canonical.name)
# 
# gpsSub <- gps[gps$Stu_Ind_Sps == dups[i],]
# gps_l <- split(gpsSub, as.character(gpsSub$tag_local_identifier))
# 
# plot(location.lat~location.long, data=gpsSub, type="n")
# cols <- c("black","red","blue","green","orange","yellow","magenta","brown")
# lapply(1:length(gps_l), function(p){
#   lines(location.lat~location.long, data=gps_l[[p]][order(gps_l[[p]]$timestamp),], col=cols[p])
#   print(range(gps_l[[p]][order(gps_l[[p]]$timestamp),"timestamp"]))
# })


#_____________________________
# ## Mark excluded studies ####
# # In the study summary table, where we notes which studies were downloaded, we mark which were excluded because of duplicates:
# 
# studies_summaryTable <- read.csv("MovementData/studies_summaryTable_downloadOrNot.csv", as.is=T)
# 
# keptStudies <- unique(referenceTableStudies_ALL$MBid[referenceTableStudies_ALL$excluded=="no"])
# duplicatedStudies <- unique(referenceTableStudies_ALL$MBid)[!unique(referenceTableStudies_ALL$MBid) %in% keptStudies]
# 
# studies_summaryTable$studyDuplicated <- "no"
# studies_summaryTable$studyDuplicated[studies_summaryTable$id %in% as.character(duplicatedStudies)] <- "yes all individual in study are duplicates"
# studies_summaryTable$studyDuplicated[studies_summaryTable$studyDownload %in% unique(grep("no ",studies_summaryTable$studyDownload,value=T))] <- NA
# 
# write.csv(studies_summaryTable, "MovementData/studies_summaryTable_downloadOrNot_duplicateOrNot.csv", row.names=F)
#
